Fixed #1!!!
The problem is that there are a set of special cases for functions that won't trigger the callback, particularly functions that start the whole process. This is how it works to call things in "test" mode in swirlify. I just had to add "exercise" to that list. There may be other ways to fix the problem.