Add: !uses_func("start")(e$expr)[[1]] && to the resume.default function in swirl.R, otherwise the lesson suddenly starts over in the middle.  Repeating Dima's explanation foudn here: https://github.com/dimagor/socraticswirl/commit/a8f809e6e9489297af0962b361321b253959de41
The problem is that there are a set of special cases for functions that won't trigger the callback, particularly functions that start the whole process. This is how it works to call things in "test" mode in swirlify. I just had to add "exercise" to that list. There may be other ways to fix the problem.